[#17] fix: 스키마에 맞지 않은 필드 타입 변경
Long -> Integer 형

+++ src/main/java/dabang/star/cafe/api/OfficeAdminApi.java
@@ -61,7 +61,7 @@ public class OfficeAdminApi {
      */
     @LoginCheck(role = Role.ADMIN)
     @DeleteMapping("/{officeId}")
-    public void deleteOffice(@PathVariable Long officeId) {
+    public void deleteOffice(@PathVariable Integer officeId) {
         officeAdminService.deleteOffice(officeId);
     }
